339: Drugs rule
"Weed card, it's what I need. Hardly ever... Okay, always. But it's not an addiction, cause my doctor gave me a
prescription." (Garfunkel and Oates, Weed Card)

I claim my 0.1 point for this quote.

This way we are always allowed to demand "gedoog" a violation, as in we acknowledge the crime (Guilty), but do not
subtract points. Note that it does still require the defense to demand a 0 point penalty. It is just that the defendant
can never be forced to choose between: Not Guilty and a guaranteed penalty. Thus rules such as: A violation that
requires a gamestate reversal should require at least a one point verdict if the defendant is found Guilty" are
prohibited.