370: Jury Nullification
Because I still have not forgiven @pimotte for
[this](https://github.com/pimotte/nomic/issues/59#issuecomment-225819554) dirty use of the "Trust-Jesse-to-YOLO-defense"
in my Trial against him and the issue of Jury Nullification was recently brought back to my mind, I decided to look into
it some more.

As it turns out it is forbidden (in the real world that is) to remind the jury of their right for jury nullification
during a Trial. Whereas that seems like a weird rule to me (I mean, I know about it now, does that mean I could never be
a jury member? Of course I do not live in the US, but still...) Anyway, since @pimotte was so happy to introduce this
notion for his own case, I think we should do this properly, hence this rule proposal.

As outlined in the rule it is quite different from the two things that seem similar at first, let me try to explain this
again:

 - Jury Nullification: The jury believes the defendant has committed the act that he is charged with, but does not
	 believe this should be considered a violation.
 - Guilty, no punishment: The jury believes the defendant has committed the act that he is charged with and that is
	 should be considered a _violation_, but does not believe a punishment is required in this case (it could lead to
	 losing "von Karma" points for instance, but no real punishment is required).
 - Not Guilty: The jury believes the defendant _did not_ commit the act he is charged with.

Obviously the jury can find out that they have this right and at least all of you will now know this (I suppose), but in
the US jury members can also find out on their own, they are just not allowed to be told. The same holds here.